cometa_ai/requirements.txt
anyio==4.4.0
arrow==1.3.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Flask==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ollama==0.5.4
python-dateutil==2.9.0.post0
redis==5.0.8
Redis-Sentinel-Url==1.0.1
rq==1.16.2
rq-dashboard==0.7.1.2
six==1.16.0
sniffio==1.3.1
types-python-dateutil==2.9.0.20240821
Werkzeug==3.0.4
playwright==1.50.0
langchain==0.3.23
nest-asyncio==1.6.0

# Django API dependencies
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg[binary]==3.2.6
PyPDF2==3.0.1
markdown==3.7.0
beautifulsoup4==4.13.3
numpy==2.2.0
chromadb==1.1.0
python-dotenv==1.1.0
onnxruntime==1.22.1
uvicorn[standard]==0.34.0
gunicorn==23.0.0
